+++ saltaway/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-from .api import *
-from .repositories import *
-from .exceptions import *
+from .api import push, pull
+from .repositories import Repository, ArchiveIs, InternetArchive
+from .exceptions import SaltawayException, RepositoryFailure, ArchivalFailure