+++ saltaway/api.py
@@ -10,10 +10,10 @@ REPOSITORIES = (
 )
 
 
-def push(url, repos=REPOSITORIES):
+def push(url, repos=REPOSITORIES, max_age=0):
     with concurrent.futures.ThreadPoolExecutor() as executor:
         future_to_repo = {
-            executor.submit(repo().push, url): repo
+            executor.submit(repo().push, url, max_age=max_age): repo
             for repo in repos
         }
         for future in concurrent.futures.as_completed(future_to_repo):